modularized note and search
