+++ bin/tv2
@@ -1,7 +1,7 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python3
 """A fast note-taking app for the UNIX terminal"""
 
-from six.moves import configparser
+import configparser
 import argparse
 import os
 import logging
@@ -48,7 +48,7 @@ the default default will be used"""
             parents=[parser],
             formatter_class=argparse.RawDescriptionHelpFormatter)
     parser.add_argument("-e", "--editor", dest="editor", action="store",
-        default=defaults.get("editor", os.getenv("EDITOR", "vim")),
+        default=defaults.get("editor", os.getenv("EDITOR", "nano")),
         help="the text editor to use (default: %(default)s)")
 
     parser.add_argument("-x", "--extension", dest="extension", action="store",
